Rename CssStyles.ts to common directory
