+++ mathjax3-ts/output/chtml/Wrappers/mtr.ts
@@ -29,7 +29,7 @@ import {CHTMLmtd} from './mtd.js';
 import {BBox} from '../BBox.js';
 import {MmlMtr, MmlMlabeledtr} from '../../../core/MmlTree/MmlNodes/mtr.js';
 import {MmlNode} from '../../../core/MmlTree/MmlNode.js';
-import {StyleList} from '../CssStyles.js';
+import {StyleList} from '../../common/CssStyles.js';
 import {DIRECTION} from '../FontData.js';
 
 /*****************************************************************/
