+++ mathjax3-ts/output/chtml/Wrappers/scriptbase.ts
@@ -30,7 +30,7 @@ import {CHTMLmo} from './mo.js';
 import {MmlMsubsup} from '../../../core/MmlTree/MmlNodes/msubsup.js';
 import {MmlNode} from '../../../core/MmlTree/MmlNode.js';
 import {BBox} from '../BBox.js';
-import {StyleData, StyleList} from '../CssStyles.js';
+import {StyleData, StyleList} from '../../common/CssStyles.js';
 import {DIRECTION} from '../FontData.js';
 
 /*
